rename generator-base to generator
