+++ generators/base-core/generator.mts
